Add `rename-default-toolbar-title-localeText` codemod
