+++ packages/x-codemod/README.md
@@ -95,6 +95,7 @@ The list includes these transformers
 - [`migrate-to-components-componentsProps`](#migrate-to-components-componentsProps)
 - [`replace-arrows-button-slot`](#replace-arrows-button-slot)
 - [`rename-should-disable-time`](#rename-should-disable-time)
+- [`rename-default-toolbar-title-localeText`](#rename-default-toolbar-title-localeText)
 
 #### `adapter-change-import`
 
@@ -339,6 +340,32 @@ Replace `shouldDisableTime` by `shouldDisableClock`.
 
 ```sh
 npx @mui/x-codemod v6.0.0/pickers/rename-should-disable-time <path>
+```
+
+#### `rename-default-toolbar-title-localeText`
+
+Rename toolbar related translation keys, removing `Default` part from them to better fit their usage.
+
+```diff
+ <LocalizationProvider
+   localeText={{
+-    datePickerDefaultToolbarTitle: 'Date Picker',
++    datePickerToolbarTitle: 'Date Picker',
+
+-    timePickerDefaultToolbarTitle: 'Time Picker',
++    timePickerToolbarTitle: 'Time Picker',
+
+-    dateTimePickerDefaultToolbarTitle: 'Date Time Picker',
++    dateTimePickerToolbarTitle: 'Date Time Picker',
+
+-    dateRangePickerDefaultToolbarTitle: 'Date Range Picker',
++    dateRangePickerToolbarTitle: 'Date Range Picker',
+   }}
+ />
+```
+
+```sh
+npx @mui/x-codemod v6.0.0/pickers/rename-default-toolbar-title-localeText <path>
 ```
 
 #### `rename-components-to-slots`
