+++ packages/x-codemod/src/v6.0.0/pickers/preset-safe/actual.spec.js
@@ -11,7 +11,13 @@ const theme = createTheme({});
 function App() {
   return (
     <ThemeProvider theme={theme}>
-      <LocalizationProvider dateAdapter={DayjsUtils} locale="fr">
+      <LocalizationProvider
+        dateAdapter={DayjsUtils}
+        locale="fr"
+        localeText={{
+          datePickerDefaultToolbarTitle: 'Pick a Date',
+        }}
+      >
         <CalendarPicker
           date={null}
           onChange={() => {}}
