+++ packages/x-codemod/src/v6.0.0/pickers/preset-safe/index.ts
@@ -8,6 +8,7 @@ import transformReplaceToolbarPropsBySlot from '../replace-toolbar-props-by-slot
 import transformMigrateToComponentsComponetsProps from '../migrate-to-components-componentsProps';
 import transformReplaceArrowsButtonSlot from '../replace-arrows-button-slot';
 import transformRenameShouldDisableTime from '../rename-should-disable-time';
+import transformRenameDefaultToolbarTitleLocaleText from '../rename-default-toolbar-title-localeText';
 
 import { JsCodeShiftAPI, JsCodeShiftFileInfo } from '../../../types';
 
@@ -25,6 +26,7 @@ export default function transformer(file: JsCodeShiftFileInfo, api: JsCodeShiftA
   file.source = transformMigrateToComponentsComponetsProps(file, api, options);
   file.source = transformReplaceArrowsButtonSlot(file, api, options);
   file.source = transformRenameShouldDisableTime(file, api, options);
+  file.source = transformRenameDefaultToolbarTitleLocaleText(file, api, options);
 
   return file.source;
 }
