+++ packages/x-codemod/src/v6.0.0/pickers/rename-default-toolbar-title-localeText/actual.spec.js
@@ -1,0 +1,11 @@
+<LocalizationProvider
+  localeText={{
+    start: 'Check-in',
+    datePickerDefaultToolbarTitle: 'Pick a Date',
+    timePickerDefaultToolbarTitle: 'Pick a Time',
+    dateTimePickerDefaultToolbarTitle: 'Pick a Date and Time',
+    dateRangePickerDefaultToolbarTitle: 'Pick a Date Range',
+  }}
+>
+  <DatePicker />
+</LocalizationProvider>;
