+++ packages/x-codemod/src/v6.0.0/pickers/rename-default-toolbar-title-localeText/expected.spec.js
@@ -1,0 +1,11 @@
+<LocalizationProvider
+  localeText={{
+    start: 'Check-in',
+    datePickerToolbarTitle: 'Pick a Date',
+    timePickerToolbarTitle: 'Pick a Time',
+    dateTimePickerToolbarTitle: 'Pick a Date and Time',
+    dateRangePickerToolbarTitle: 'Pick a Date Range',
+  }}
+>
+  <DatePicker />
+</LocalizationProvider>;
