+++ packages/x-codemod/src/v6.0.0/pickers/replace-toolbar-props-by-slot/actual.spec.js
@@ -18,15 +18,3 @@
 <DatePicker showToolbar={isToolbarVisible(x, y)} />;
 
 <DatePicker toolbarTitle="Title" />;
-
-<LocalizationProvider
-  localeText={{
-    start: 'Check-in',
-    datePickerDefaultToolbarTitle: 'Pick a Date',
-    timePickerDefaultToolbarTitle: 'Pick a Time',
-    dateTimePickerDefaultToolbarTitle: 'Pick a Date and Time',
-    dateRangePickerDefaultToolbarTitle: 'Pick a Date Range',
-  }}
->
-  <DatePicker />
-</LocalizationProvider>;
