+++ packages/x-codemod/src/v6.0.0/pickers/replace-toolbar-props-by-slot/expected.spec.js
@@ -43,15 +43,3 @@
   localeText={{
     toolbarTitle: "Title",
   }} />;
-
-<LocalizationProvider
-  localeText={{
-    start: 'Check-in',
-    datePickerToolbarTitle: 'Pick a Date',
-    timePickerToolbarTitle: 'Pick a Time',
-    dateTimePickerToolbarTitle: 'Pick a Date and Time',
-    dateRangePickerToolbarTitle: 'Pick a Date Range',
-  }}
->
-  <DatePicker />
-</LocalizationProvider>;
