+++ packages/x-codemod/src/v6.0.0/pickers/replace-toolbar-props-by-slot/index.ts
@@ -1,7 +1,6 @@
 import type { JsCodeShiftAPI, JsCodeShiftFileInfo } from '../../../types';
 import { transformNestedProp } from '../../../util/addComponentsSlots';
 import removeProps from '../../../util/removeProps';
-import renameNestedProps from '../../../util/renameNestedProps';
 
 const propsToSlots = {
   ToolbarComponent: { prop: 'components', path: 'Toolbar' },
@@ -9,15 +8,6 @@ const propsToSlots = {
   toolbarFormat: { prop: 'componentsProps', path: 'toolbar.toolbarFormat' },
   showToolbar: { prop: 'componentsProps', path: 'toolbar.hidden' },
   toolbarTitle: { prop: 'localeText', path: 'toolbarTitle' },
-};
-
-const propsToRename = {
-  localeText: {
-    datePickerDefaultToolbarTitle: 'datePickerToolbarTitle',
-    timePickerDefaultToolbarTitle: 'timePickerToolbarTitle',
-    dateTimePickerDefaultToolbarTitle: 'dateTimePickerToolbarTitle',
-    dateRangePickerDefaultToolbarTitle: 'dateRangePickerToolbarTitle',
-  },
 };
 
 const COMPONENTS = [
@@ -94,13 +84,6 @@ export default function transformer(file: JsCodeShiftFileInfo, api: JsCodeShiftA
     });
 
   removeProps({ root, componentNames: COMPONENTS, props: Object.keys(propsToSlots), j });
-
-  renameNestedProps({
-    root,
-    componentNames: ['LocalizationProvider'],
-    nestedProps: propsToRename,
-    j,
-  });
 
   return root.toSource(printOptions);
 }
